+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file. This change
  - Logging about retries during connections
  - Request :guild-members-chunk events
  - Ability to update bot's status on Discord
+ - Ability to update bot's voice state on Discord
 
 ### Removed
  - Uberjar profile in Leiningen